Removed Base class and added functionality with defining graphs and gradients
