+++ Model/backend.py
@@ -813,75 +813,3 @@ def progress_bar(progress, maximum):
     # update the progress bar in the terminal
     sys.stdout.write('\r[%s] %s%s' % (bar, percent, '%'))
     # sys.stdout.flush()
-
-
-class Base:
-    """ to make the Simulation class a bit cleaner and less confusing
-        this class is a parent to the Simulation class to hid methods
-        that shouldn't need to be modified
-    """
-    def __init__(self):
-        self.number_cells = 0
-        self.cell_array_names = list()
-        self.array_types = dict()
-        self.function_times = dict()
-
-    def cell_type(self, name, number):
-        """ creates a new cell type for setting initial parameters
-            and defines a slice in the cell arrays that corresponds
-            to this cell type
-        """
-        # determine the bounds of the slice and update the number of cells
-        begin = self.number_cells
-        end = self.number_cells = begin + number
-
-        # add slice to general dictionary for
-        self.array_types[name] = (begin, end)
-
-        for graph_name in self.graph_names:
-            self.__dict__[graph_name].add_vertices(number)
-
-    def cell_arrays(self, *args):
-        """ creates Simulation instance variables for the cell arrays
-            used to represent cell values
-        """
-        # go through all array parameters provided
-        for array_params in args:
-            # add the array name to a list for automatic addition/removal when cells divide and die
-            self.cell_array_names.append(array_params[0])
-
-            # get the tuple length for the particular array parameters
-            length = len(array_params)
-
-            # get the initial size of 1D array if the length is 2, but make a 2D array if a third index is provided
-            if length == 2:
-                size = self.number_cells
-
-            elif length == 3:
-                size = (self.number_cells, array_params[2])
-
-            else:
-                raise Exception("Tuples for defining cell array parameters should have length 2 or 3.")
-
-            # create an instance variable of the Simulation class for the cell array with these parameters
-            self.__dict__[array_params[0]] = np.empty(size, dtype=array_params[1])
-
-    def initials(self, array_name, func, cell_type=None):
-        """ given a lambda function for the initial values
-            of a cell array this updates that accordingly
-        """
-        # if no cell type name provided
-        if cell_type is None:
-            # go through all cells and give this initial parameter
-            for i in range(self.number_cells):
-                self.__dict__[array_name][i] = func()
-
-        # otherwise update the slice of the array based on the
-        else:
-            # get the beginning and end of the slice
-            begin = self.array_types[cell_type][0]
-            end = self.array_types[cell_type][1]
-
-            # update only this slice of the cell array
-            for i in range(begin, end):
-                self.__dict__[array_name][i] = func()
